Changes to be committed: 	modified:   .amake/cppSizes 	modified:   build/bin/NanoX 	modified:   build/obj/chars.o 	modified:   build/obj/cut.o 	modified:   build/obj/files.o 	modified:   build/obj/global.o 	modified:   build/obj/nano.o 	modified:   build/obj/text.o 	modified:   build/obj/winio.o 	modified:   src/cpp/cut.cpp 	modified:   src/cpp/files.cpp 	modified:   src/cpp/global.cpp 	modified:   src/cpp/help.cpp 	modified:   src/cpp/nano.cpp 	modified:   src/cpp/text.cpp 	modified:   src/cpp/winio.cpp 	modified:   src/include/constexpr_def.h 	modified:   src/include/definitions.h 	modified:   src/include/prototypes.h

+++ src/cpp/help.cpp
@@ -515,7 +515,7 @@ show_help()
     //  The current answer when the user invokes help at the prompt.
     //
     s8 *saved_answer = (answer != nullptr) ? copy_of(answer) : nullptr;
-    u32 stash[sizeof(flags) / sizeof(flags[0])];
+    u64 stash[sizeof(flags) / sizeof(flags[0])];
 
     //
     //  A storage place for the current flag settings.
